+++ sync-for-payables/src/sdk/models/operations/createbillcreditnote.ts
@@ -58,5 +58,5 @@ export class CreateBillCreditNoteResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }