+++ sync-for-payables/src/sdk/models/operations/downloadbillattachment.ts
@@ -61,5 +61,5 @@ export class DownloadBillAttachmentResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }