+++ sync-for-payables/src/sdk/models/operations/listbillcreditnotes.ts
@@ -67,5 +67,5 @@ export class ListBillCreditNotesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }