+++ sync-for-payables/src/sdk/models/operations/listtaxrates.ts
@@ -61,7 +61,7 @@ export class ListTaxRatesResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Success